requirements.txt
fastapi
uvicorn
pydantic
alpha_vantage
python-dotenv
pandas_ta
transformers
torch
tf-keras
finnhub-python
newsapi-python